Add support and tests to cover invalid URLs and URLs with whitespaces.
If a get request is attempted to the UrlBox API with an invalid URL then
an InvalidUrlException is raised, before the call is made to the API.

This allows the user to deal with the invalid URL themselves.

This uses the validator package to validate URLs.

If a URL with whitespace before or after the URL is passed in to the get
request, these are stripped before validation, encoding and finally the
get request is made.

refs:
https://validators.readthedocs.io/en/latest/