Add slightly stricter regex on validating sha256 signature.
Not all of the webhook_validator was covered by tests.

The check for an invalid sha256 signature was set to pending as we
need to implement the more thorough server generated sha256 signature
vs the client generated sha256 signature.

However this pending test was the only test testing the raising of the
invalid signature exception:

if not re.search(SIGNATURE_REGEX, raw_signature):
    raise (InvalidHeaderSignatureError("Invalid signature"))

By making the regex checking the sha256 signature stricter, by requiring
at least 40 characters in the signature, we can now test this area of the
code, rather than have the test set to pending.