Refactor urlbox_client.py
This extracts out the processing for the options into its own private
method, "_process_options".

This performs the nec. checks for required keys, handles default values
if format is not supplied, processes the url and finally url encodes the
options.

It returns the processed, encoded options and the format as a tuple.