Do not url encode options when making a get request.
This is something taken from the ruby gem version of this library.

The need to play around with json loads and dumps was an attempt to
deal with the url encoded string of options being marshalled into a json
array for the post request body.

Now we are using the options dictionary as is, it doesn't need the json
marshalling anymore.

Somewhere along the line the post function broke due to the json
marshalling.